pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scholarforge"
version = "0.1.0"
description = "ScholarForge - AI-powered research and report generation platform"
readme = "readme.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "ScholarForge Team"}
]

dependencies = [
    "celery[redis]>=5.3.0",
    "redis>=4.5.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "google-search-results>=2.4.2",
    "python-docx>=0.8.11",
    "reportlab>=4.0.0",
    "sqlalchemy>=2.0.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "pymupdf>=1.23.0",
    "lxml>=4.9.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "itsdangerous>=2.1.0",
    "psycopg2-binary>=2.9.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "watchdog>=3.0.0",
    "pypandoc>=1.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools]
packages = ["backend"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
